pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "flask>=3.1.2",
    "flask-cors>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "pyinstaller>=6.15.0",
    "python-dotenv>=1.1.1",
]
